updated buildgradle

+++ code/js/src/service/mock/UserServiceMock.ts
@@ -3,6 +3,7 @@ import {User} from "../../domain/User";
 import {UserService} from "../interfaces/UserService";
 import {delay} from "./utils";
 import {Repo} from "../../App";
+import { tokenHandler } from "./tokenHandler";
 
 export class UserServiceMock implements UserService {
     repo: Repo;
@@ -10,29 +11,8 @@ export class UserServiceMock implements UserService {
     constructor(repo: Repo) {
         this.repo = repo;
     }
-    login(username: string, password: string): Promise<User> {
-        throw new Error("Method not implemented.");
-    }
-    register(email: string, username: string, password: string, code: string): Promise<User> {
-        throw new Error("Method not implemented.");
-    }
-    registerFirstUser(email: string, username: string, password: string): Promise<User> {
-        throw new Error("Method not implemented.");
-    }
-    logOut(): Promise<void> {
-        throw new Error("Method not implemented.");
-    }
-    updateUsername(newUsername: string): Promise<User> {
-        throw new Error("Method not implemented.");
-    }
-    getUserById(userId: number): Promise<User> {
-        throw new Error("Method not implemented.");
-    }
-    searchByUsername(username: string, limit?: number, skip?: number): Promise<User[]> {
-        throw new Error("Method not implemented.");
-    }
-}
-  /*  async login(username: string, password: string): Promise<User | undefined> {
+    
+    async login(username: string, password: string): Promise<User | undefined> {
         await delay(500);
         const user = this.repo.userRepo.users.find(user => user.username === username);
         if (!user) {
@@ -43,12 +23,16 @@ export class UserServiceMock implements UserService {
             return undefined;
         }
         const token = this.repo.userRepo.createToken(user.id);
+        tokenHandler().setToken(token);
         console.log(this.repo.userRepo.tokens);
         return Promise.resolve(user)
     }
 
     async register(email: string, username:string, password: string, code:string): Promise<User> {
         await delay(1000);
+        const token = tokenHandler().getToken();
+        if(!token) throw new Error("Invalid token");
+        if(!this.repo.userRepo.getUserByToken(token)) throw new Error("Invalid token");
         //TODO check if email is already in use
         if (this.repo.userRepo.users.find(user => user.email === email)) {
             throw new Error("Email already in use");
@@ -60,13 +44,19 @@ export class UserServiceMock implements UserService {
         return this.repo.userRepo.createUser(username, email, password);
     }
 
-    async logOut(token: string): Promise<void> {
+    async logOut(): Promise<void> {
         await delay(1000);
+        const token = tokenHandler().getToken();
+        if(!token) throw new Error("Invalid token");
+        if(!this.repo.userRepo.getUserByToken(token)) throw new Error("Invalid token");
         this.repo.userRepo.deleteToken(token);
     }
 
-    async updateUsername(token: string, newUsername: string): Promise<User> {
+    async updateUsername(newUsername: string): Promise<User> {
         await delay(1000);
+        const token = tokenHandler().getToken();
+        if(!token) throw new Error("Invalid token");
+        if(!this.repo.userRepo.getUserByToken(token)) throw new Error("Invalid token");
         const user = this.repo.userRepo.getUserByToken(token);
         if (!user) {
             throw new Error("User not found");
@@ -79,8 +69,11 @@ export class UserServiceMock implements UserService {
         return user;
     }
 
-    async getUserById(token: string, userId: number): Promise<User> {
+    async getUserById(userId: number): Promise<User> {
         await delay(500);
+        const token = tokenHandler().getToken();
+        if(!token) throw new Error("Invalid token");
+        if(!this.repo.userRepo.getUserByToken(token)) throw new Error("Invalid token");
         const userVerify = this.repo.userRepo.getUserByToken(token);
         if (!userVerify) {
             throw new Error("Invalid token");
@@ -92,8 +85,11 @@ export class UserServiceMock implements UserService {
         return user;
     }
 
-    async searchByUsername(token: string, username: string, limit: number = 10, skip: number = 0): Promise<User[]> {
+    async searchByUsername(username: string, limit: number = 10, skip: number = 0): Promise<User[]> {
         await delay(500);
+        const token = tokenHandler().getToken();
+        if(!token) throw new Error("Invalid token");
+        if(!this.repo.userRepo.getUserByToken(token)) throw new Error("Invalid token");
         const userVerify = this.repo.userRepo.getUserByToken(token);
         if (!userVerify) {
             throw new Error("Invalid token");
@@ -111,5 +107,5 @@ export class UserServiceMock implements UserService {
         }
         return this.repo.userRepo.createUser(username, email, password);
     }
-}*/
+}
